refactor(types); extract public interfaces for VData*
