+++ packages/vuetify/src/components/VDataTable/VDataTableHeaderDesktop.ts
@@ -1,8 +1,9 @@
 // Helpers
 import { VNode } from 'vue'
 import mixins from '../../util/mixins'
-import header, { TableHeader } from './mixins/header'
+import header from './mixins/header'
 import { wrapInArray, convertToUnit } from '../../util/helpers'
+import { TableHeader } from './types'
 
 export default mixins(header).extend({
   name: 'v-data-table-header-desktop',
